add: nao sei o que funciona

+++ src/main/java/com/fitmind/repository/HealthProfileRepository.java
@@ -4,8 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.fitmind.entity.HealthProfile;
 
-public interface HealthProfileRepository extends JpaRepository<HealthProfile, Integer> {
+public interface HealthProfileRepository extends JpaRepository<HealthProfile, Long> {
 	
-	public HealthProfile findByUserId(Integer id);
+	public HealthProfile findByUserId(Long id);
 
 }
